updated stack creation to prevent overflows

+++ include/push_swap.h
@@ -16,6 +16,7 @@
 # include <stdlib.h>
 # include <stdbool.h>
 # include <stdio.h>
+# include <limits.h>
 
 typedef struct s_list
 {
@@ -38,7 +39,7 @@ bool	validate_args(char **args, int start);
 // Utils
 t_split	ft_split(char const *s, char c);
 size_t	ft_strlen(const char *str);
-int		ft_atoi(const char *str);
+long	ft_atoi(const char *str);
 char	*ft_substr(char const *s, unsigned int start, size_t len);
 char	*ft_strdup(const char *src);
 void	list_clear(t_list *list);
